safari-600.7-branch/LayoutTests/mathml/presentation/menclose-notation-no-overlap-expected.txt

Tests that menclose with notation value as circle is sqrt(2) times the inner width/height.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('mathmenclose1').offsetWidth >= 141 is true
PASS document.getElementById('mathmenclose1').offsetHeight >= 70 is true
PASS successfullyParsed is true

TEST COMPLETE
